refactor move useInstallableViewItemactions
